Restructure node_report.cc, move functions to utilities.cc
